loan_API/requirements.txt
aiosmtplib==3.0.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asg==1.1.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.1
blinker==1.9.0
catboost==1.2.7
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
constantly==23.10.4
contourpy==1.0.7
cryptography==44.0.1
cycler==0.12.1
daphne==4.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-mail==1.4.2
fonttools==4.56.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
narwhals==1.27.1
numba==0.61.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
plotly==6.0.0
pluggy==1.5.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
redis==5.2.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
service-identity==24.2.0
setuptools==75.8.1
shap==0.46.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.22
sqlparse==0.5.3
starlette==0.45.3
threadpoolctl==3.5.0
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
zope.interface==7.2
pyodbc

